Add files via upload
The three code files I wrote. 
Rules_Analysis and Alpha Project RNotebook are both R notebooks. Alpha Project RNotebook is the main piece of code in which I pre-processed the data, wrote the function that mines hyperedge sets and saves it all as .txt files on the server. Rules_Analysis is the piece of code that I wrote after I had finished mining the rules, in which I separated the trivial rules from non-trivial rules.
RscriptUpdated is the code I ran on the server to mine the association rules. In general, I took the relevant code from Alpha Project RNotebook and put it all in a single block of code that the server could run properly.